Web: move notification actions to shared file
